[Store][Refactor] 쿼리dsl에서 vector 함수 쓰도록
- double[]은 지원 안 함 float[]만 지원함

+++ src/main/java/com/bobeat/backend/domain/store/external/clova/dto/ClovaEmbeddingResponse.java
@@ -20,6 +20,6 @@ public class ClovaEmbeddingResponse {
     @AllArgsConstructor
     @JsonIgnoreProperties(ignoreUnknown = true)
     public static class Result {
-        private List<Double> embedding;
+        private List<Float> embedding;
     }
 }
